Linter changings, deleted keyboard after push
